backend/requirements.txt
Flask==3.0.2
face-recognition==1.3.0
opencv-python==4.9.0.80
mysql-connector-python==8.3.0
requests==2.31.0
numpy==1.26.4